style(dropdown-web): split dropdown and menu
