+++ packages/pluggableWidgets/dropdown-web/src/Dropdown.editorPreview.tsx
@@ -1,7 +1,7 @@
 // import { parseStyle } from "@mendix/pluggable-widgets-commons";
 import { createElement, ReactElement } from "react";
 
-import Dropdown from "./Dropdown";
+import { Dropdown } from "./components/Dropdown";
 
 export const preview = (): ReactElement => {
     // TODO: Change PIW preview props typing (class -> className) generation to remove the ts-ignore below
